Configurando Permisos Administrativos

+++ application/controllers/Administrador/Permisos_Controller.php
@@ -11,15 +11,17 @@ class Permisos_Controller extends CI_Controller
         if (!$this->session->userdata("login")) {
             redirect(base_url());
         }
-
         $this->load->model("permisos_model");
         $this->load->model("usuarios_model");
+        $this->accesos = $this->backend_lib->control();
     }
 
     public function index()
     {
+        if(! $this->accesos->p_read){ redirect(base_url()); return; }
         $data = array(
             'permisos' => $this->permisos_model->getPermisos(),
+            'accesos' => $this->accesos,
         );
         $this->load->view('layouts/header');
         $this->load->view('layouts/aside');
@@ -43,12 +45,14 @@ class Permisos_Controller extends CI_Controller
 
     public function delete($id)
     {
+        if(! $this->accesos->p_delete){ redirect(base_url()); return; }
         $this->permisos_model->delete($id);
         redirect(base_url() . "administrador/Permisos_Controller");
     }
 
     public function edit($id)
     {
+        if(! $this->accesos->p_update){ redirect(base_url()); return; }
         $data  = array(
             'roles' => $this->usuarios_model->getRoles(),
             'menus' => $this->permisos_model->getMenus(),
